feat(allRoutes): Create token verification system to secured the app

+++ backend/routes/tips.ts
@@ -1,6 +1,7 @@
 import { PrismaClient } from '@prisma/client';
 import dotenv from "dotenv";
 import express from 'express';
+import { checkIfTokenIsValid, isManager, getEmployeeIdFromToken } from '../controllers/tokenCheck';
 
 const router = express.Router();
 const prisma = new PrismaClient();
@@ -8,6 +9,16 @@ dotenv.config();
 
 router.get('/', async (req, res) => {
     try {
+        const authHeader = req.headers.authorization;
+        if (!authHeader) {
+            res.status(401).send("Authorization header missing");
+            return;
+        }
+        const token = authHeader.split(" ")[1];
+        if (!await checkIfTokenIsValid(token)) {
+            res.status(401).send("Invalid token");
+            return;
+        }
         const tips = await prisma.tips.findMany();
         res.send(tips);
     } catch (error) {
@@ -17,10 +28,28 @@ router.get('/', async (req, res) => {
 
 router.get('/:id', async (req, res) => {
     try {
-        const { id } = req.params;
+        const authHeader = req.headers.authorization;
+        if (!authHeader) {
+            res.status(401).send("Authorization header missing");
+            return;
+        }
+        const token = authHeader.split(" ")[1];
+        if (!await checkIfTokenIsValid(token)) {
+            res.status(401).send("Invalid token");
+            return;
+        }
+        if (req.params.id === null) {
+            res.status(400).send('ID missing');
+            return;
+        }
+        if (isNaN(parseInt(req.params.id))) {
+            res.status(400).send('ID must be a number');
+            return
+        }
+        const id = parseInt(req.params.id);
         const tip = await prisma.tips.findUnique({
             where: {
-                id: parseInt(id)
+                id: id
             }
         });
         if (!tip) {
@@ -35,6 +64,24 @@ router.get('/:id', async (req, res) => {
 
 router.post('/', async (req, res) => {
     try {
+        const authHeader = req.headers.authorization;
+        if (!authHeader) {
+            res.status(401).send("Authorization header missing");
+            return;
+        }
+        const token = authHeader.split(" ")[1];
+        if (!await checkIfTokenIsValid(token)) {
+            res.status(401).send("Invalid token");
+            return;
+        }
+        if (!await isManager(token)) {
+            res.status(401).send("Unauthorized");
+            return;
+        }
+        if (!req.body.title || !req.body.tip) {
+            res.status(400).send('Title and tip are required');
+            return;
+        }
         const maxId = await prisma.tips.findFirst({
             select: {
                 id: true
@@ -59,11 +106,37 @@ router.post('/', async (req, res) => {
 
 router.put('/:id', async (req, res) => {
     try {
-        const { id } = req.params;
+        const authHeader = req.headers.authorization;
+        if (!authHeader) {
+            res.status(401).send("Authorization header missing");
+            return;
+        }
+        const token = authHeader.split(" ")[1];
+        if (!await checkIfTokenIsValid(token)) {
+            res.status(401).send("Invalid token");
+            return;
+        }
+        if (!await isManager(token)) {
+            res.status(401).send("Unauthorized");
+            return;
+        }
+        if (req.params.id === null) {
+            res.status(400).send('ID missing');
+            return;
+        }
+        if (isNaN(parseInt(req.params.id))) {
+            res.status(400).send('ID must be a number');
+            return
+        }
+        if (!req.body.title || !req.body.tip) {
+            res.status(400).send('Title and tip are required');
+            return;
+        }
+        const id = parseInt(req.params.id);
         const { newTitle, newTip } = req.body;
         const updatedTip = await prisma.tips.update({
             where: {
-                id: parseInt(id)
+                id: id
             },
             data: {
                 title: newTitle,
@@ -82,10 +155,32 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
     try {
-        const { id } = req.params;
+        const authHeader = req.headers.authorization;
+        if (!authHeader) {
+            res.status(401).send("Authorization header missing");
+            return;
+        }
+        const token = authHeader.split(" ")[1];
+        if (!await checkIfTokenIsValid(token)) {
+            res.status(401).send("Invalid token");
+            return;
+        }
+        if (!await isManager(token)) {
+            res.status(401).send("Unauthorized");
+            return;
+        }
+        if (req.params.id === null) {
+            res.status(400).send('ID missing');
+            return;
+        }
+        if (isNaN(parseInt(req.params.id))) {
+            res.status(400).send('ID must be a number');
+            return
+        }
+        const id = parseInt(req.params.id);
         const deletedTip = await prisma.tips.delete({
             where: {
-                id: parseInt(id)
+                id: id
             }
         });
         if (!deletedTip) {
